moving some more ops
